refactor tests, create new file WidgetSteps
